Option to enter a ssh passphrase in the terminal
Add settings enable-ssh-askpass-require and disable-ssh-askpass-require.

When enable-ssh-askpass-require is true, you will get a prompt for your
ssh passphrase on the command line.

The default behaviour is to open an X11 window to read this passphrase.

As per the ssh-add
[manpage](https://www.man7.org/linux/man-pages/man1/ssh-add.1.html#ENVIRONMENT):

If ssh-add does not have a terminal associated with it but
DISPLAY and SSH_ASKPASS are set, it will execute the
program specified by SSH_ASKPASS (by default “ssh-askpass”)
and open an X11 window to read the passphrase.

Update README with your zqs settings

Closes #192

Signed-off-by: Darren <10267761+dglynn@users.noreply.github.com>